pyproject.toml

[tool.poetry]
name = "gmail-rules-engine"
version = "0.1.0"
description = "A Python script that integrates with Gmail API and performs rule-based operations on emails"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
google-auth-oauthlib = "^1.0.0"
google-api-python-client = "^2.100.0"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
schedule = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.6.1"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"